iox-#1332 Separate implementation from .hpp into .cpp or .inl
