+++ iceoryx_examples/automotive_soa/CMakeLists.txt
@@ -27,7 +27,11 @@ include(IceoryxPlatform)
 
 # @todo #1332 enable 1:N communication cmake switch
 
-add_executable(iox-cpp-automotive-skeleton ./iox_automotive_skeleton.cpp ./minimal_skeleton.cpp)
+add_executable(iox-cpp-automotive-skeleton  ./iox_automotive_skeleton.cpp
+                                            ./minimal_skeleton.cpp
+                                            ./src/runtime.cpp
+                                            ./src/owl/kom/method_server.cpp
+)
 target_link_libraries(iox-cpp-automotive-skeleton
     iceoryx_posh::iceoryx_posh
 )
@@ -35,7 +39,11 @@ target_compile_options(iox-cpp-automotive-skeleton PRIVATE ${ICEORYX_WARNINGS} $
 
 target_include_directories(iox-cpp-automotive-skeleton PRIVATE include)
 
-add_executable(iox-cpp-automotive-proxy ./iox_automotive_proxy.cpp ./minimal_proxy.cpp)
+add_executable(iox-cpp-automotive-proxy ./iox_automotive_proxy.cpp
+                                        ./minimal_proxy.cpp
+                                        ./src/runtime.cpp
+                                        ./src/owl/kom/method_client.cpp
+)
 target_link_libraries(iox-cpp-automotive-proxy
     iceoryx_posh::iceoryx_posh
 )
