+++ iceoryx_examples/automotive_soa/include/owl/kom/field_subscriber.hpp
@@ -23,7 +23,6 @@
 #include "owl/types.hpp"
 
 #include <limits>
-#include <memory>
 
 namespace owl
 {
@@ -37,22 +36,8 @@ class FieldSubscriber
     static constexpr uint64_t QUEUE_CAPACITY{1U};
     static constexpr uint64_t HISTORY_REQUEST{1U};
 
-    FieldSubscriber(const core::String& service, const core::String& instance, const core::String& event) noexcept
-        : m_subscriber({service, instance, event}, {QUEUE_CAPACITY, HISTORY_REQUEST})
-        , m_client({service, instance, event})
-    {
-        m_waitset.attachState(m_client, iox::popo::ClientState::HAS_RESPONSE).or_else([](auto) {
-            std::cerr << "failed to attach client" << std::endl;
-            std::exit(EXIT_FAILURE);
-        });
-    }
-
-    ~FieldSubscriber()
-    {
-        m_waitset.detachState(m_client, iox::popo::ClientState::HAS_RESPONSE);
-        // Wait here if a callback is still running
-        std::lock_guard<iox::posix::mutex> guard(m_mutex);
-    }
+    FieldSubscriber(const core::String& service, const core::String& instance, const core::String& event) noexcept;
+    ~FieldSubscriber() noexcept;
 
     FieldSubscriber(const FieldSubscriber&) = delete;
     FieldSubscriber(FieldSubscriber&&) = delete;
@@ -61,118 +46,14 @@ class FieldSubscriber
 
     template <typename Callable>
     owl::core::Result<size_t> GetNewSamples(Callable&& callable,
-                                            size_t maxNumberOfSamples = std::numeric_limits<size_t>::max()) noexcept
-    {
-        IOX_DISCARD_RESULT(maxNumberOfSamples);
+                                            size_t maxNumberOfSamples = std::numeric_limits<size_t>::max()) noexcept;
 
-        owl::core::Result<size_t> numberOfSamples{0};
-
-        while (m_subscriber.take()
-                   .and_then([&](const auto& sample) {
-                       callable(sample.get());
-                       numberOfSamples++;
-                   })
-                   .or_else([](auto& result) {
-                       if (result != iox::popo::ChunkReceiveResult::NO_CHUNK_AVAILABLE)
-                       {
-                           std::cerr << "Error receiving chunk!" << std::endl;
-                       }
-                   }))
-        {
-        }
-        return numberOfSamples;
-    }
-
-    Future<FieldType> Get()
-    {
-        // If we call Get() twice shortly after each other, once the response of the first request has not yet
-        // arrived, we have a problem
-        bool requestSuccessfullySend{false};
-        m_client.loan()
-            .and_then([&](auto& request) {
-                request.getRequestHeader().setSequenceId(m_sequenceId);
-                request.send().and_then([&]() { requestSuccessfullySend = true; }).or_else([](auto& error) {
-                    std::cerr << "Could not send Request! Error: " << error << std::endl;
-                });
-            })
-            .or_else([&](auto& error) {
-                std::cerr << "Could not allocate Request! Error: " << error << std::endl;
-                requestSuccessfullySend = false;
-            });
-
-        if (!requestSuccessfullySend)
-        {
-            return Future<FieldType>();
-        }
-        return receiveResponse();
-    }
-
-    Future<FieldType> Set(const FieldType& value)
-    {
-        // If we call Set() twice shortly after each other, once the response of the first request has not yet
-        // arrived, we have a problem
-        bool requestSuccessfullySend{false};
-        m_client.loan()
-            .and_then([&](auto& request) {
-                request.getRequestHeader().setSequenceId(m_sequenceId);
-                request->emplace(value);
-                request.send().and_then([&]() { requestSuccessfullySend = true; }).or_else([](auto& error) {
-                    std::cerr << "Could not send Request! Error: " << error << std::endl;
-                });
-            })
-            .or_else([&](auto& error) {
-                std::cerr << "Could not allocate Request! Error: " << error << std::endl;
-                requestSuccessfullySend = false;
-            });
-
-        if (!requestSuccessfullySend)
-        {
-            return Future<FieldType>();
-        }
-        return receiveResponse();
-    }
+    Future<FieldType> Get() noexcept;
+    Future<FieldType> Set(const FieldType& value) noexcept;
 
   private:
-    Future<FieldType> receiveResponse()
-    {
-        Promise<FieldType> promise;
-        auto future = promise.get_future();
-        // Typically you would e.g. use a worker pool here, for simplicity we use a plain thread
-        std::thread(
-            [&](Promise<FieldType>&& promise) {
-                // Avoid race if MethodClient d'tor is called while this thread is still running
-                std::lock_guard<iox::posix::mutex> guard(m_mutex);
+    Future<FieldType> receiveResponse() noexcept;
 
-                auto notificationVector = m_waitset.timedWait(iox::units::Duration::fromSeconds(5));
-
-                for (auto& notification : notificationVector)
-                {
-                    if (notification->doesOriginateFrom(&m_client))
-                    {
-                        while (m_client.take().and_then([&](const auto& response) {
-                            auto receivedSequenceId = response.getResponseHeader().getSequenceId();
-                            if (receivedSequenceId == m_sequenceId)
-                            {
-                                FieldType result = *response;
-                                m_sequenceId++;
-                                promise.set_value_at_thread_exit(result);
-                            }
-                            else
-                            {
-                                std::cerr << "Got Response with outdated sequence ID! Expected = " << m_sequenceId
-                                          << "; Actual = " << receivedSequenceId << "!" << std::endl;
-                                std::terminate();
-                            }
-                        }))
-                        {
-                        }
-                    }
-                }
-            },
-            std::move(promise))
-            .detach();
-        return future;
-    }
     iox::popo::Subscriber<FieldType> m_subscriber;
     iox::popo::Client<iox::cxx::optional<FieldType>, FieldType> m_client;
     std::atomic<int64_t> m_sequenceId{0};
@@ -183,5 +64,7 @@ class FieldSubscriber
 
 } // namespace kom
 } // namespace owl
+
+#include "owl/kom/field_subscriber.inl"
 
 #endif // IOX_EXAMPLES_AUTOMOTIVE_SOA_FIELD_SUBSCRIBER_HPP
