+++ iceoryx_examples/automotive_soa/include/owl/runtime.hpp
@@ -38,149 +38,25 @@ class Runtime
                                          NumberOfAvailableServicesOnLastSearch>;
 
   public:
-    static Runtime& GetInstance(const core::String& name) noexcept
-    {
-        iox::runtime::PoshRuntime::initRuntime(iox::RuntimeName_t(iox::cxx::TruncateToCapacity, name));
-        static Runtime runtime;
-        return runtime;
-    }
+    static Runtime& GetInstance(const core::String& name) noexcept;
 
-    static Runtime& GetInstance() noexcept
-    {
-        static Runtime runtime;
-        return runtime;
-    }
+    static Runtime& GetInstance() noexcept;
 
     kom::ServiceHandleContainer<kom::ProxyHandleType> FindService(kom::ServiceIdentifier& serviceIdentifier,
-                                                                  kom::InstanceIdentifier& instanceIdentifier) noexcept
-    {
-        kom::ServiceHandleContainer<kom::ProxyHandleType> iceoryxServiceContainer;
-
-        m_discovery.findService(
-            serviceIdentifier,
-            instanceIdentifier,
-            iox::cxx::nullopt,
-            [&](auto& service) {
-                iceoryxServiceContainer.push_back({service.getEventIDString(), service.getInstanceIDString()});
-            },
-            iox::popo::MessagingPattern::PUB_SUB);
-
-        m_discovery.findService(
-            serviceIdentifier,
-            instanceIdentifier,
-            iox::cxx::nullopt,
-            [&](auto& service) {
-                iceoryxServiceContainer.push_back({service.getEventIDString(), service.getInstanceIDString()});
-            },
-            iox::popo::MessagingPattern::REQ_RES);
-
-        // We need to make sure that all three internal services representing 'MinimalSkeleton' are available
-        kom::ServiceHandleContainer<kom::ProxyHandleType> autosarServiceContainer;
-        if (verifyThatServiceIsComplete(iceoryxServiceContainer))
-        {
-            autosarServiceContainer.push_back({serviceIdentifier, instanceIdentifier});
-        }
-
-        return autosarServiceContainer;
-    }
+                                                                  kom::InstanceIdentifier& instanceIdentifier) noexcept;
 
     kom::FindServiceHandle StartFindService(kom::FindServiceHandler<kom::ProxyHandleType> handler,
                                             kom::ServiceIdentifier& serviceIdentifier,
-                                            kom::InstanceIdentifier& instanceIdentifier) noexcept
-    {
-        // Duplicate entries for the same service are allowed
-        m_callbacks.push_back(CallbackEntryType(
-            handler, {serviceIdentifier, instanceIdentifier}, NumberOfAvailableServicesOnLastSearch()));
+                                            kom::InstanceIdentifier& instanceIdentifier) noexcept;
 
-        if (m_callbacks.size() == 1)
-        {
-            auto invoker = iox::popo::createNotificationCallback(invokeCallback, *this);
-            m_listener.attachEvent(m_discovery, iox::runtime::ServiceDiscoveryEvent::SERVICE_REGISTRY_CHANGED, invoker)
-                .or_else([](auto) {
-                    std::cerr << "unable to attach discovery" << std::endl;
-                    std::exit(EXIT_FAILURE);
-                });
-        }
-
-        return kom::FindServiceHandle({serviceIdentifier, instanceIdentifier});
-    }
-
-    void StopFindService(kom::FindServiceHandle handle) noexcept
-    {
-        auto iter = m_callbacks.begin();
-        for (; iter != m_callbacks.end(); iter++)
-        {
-            if (std::get<1>(*iter).GetServiceId() == handle.GetServiceId()
-                && std::get<1>(*iter).GetInstanceId() == handle.GetInstanceId())
-            {
-                break;
-            }
-        }
-        if (iter != m_callbacks.end())
-        {
-            m_callbacks.erase(iter);
-        }
-
-        if (m_callbacks.empty())
-        {
-            m_listener.detachEvent(m_discovery, iox::runtime::ServiceDiscoveryEvent::SERVICE_REGISTRY_CHANGED);
-        }
-    }
+    void StopFindService(kom::FindServiceHandle handle) noexcept;
 
   private:
     explicit Runtime() noexcept = default;
 
-    bool verifyThatServiceIsComplete(kom::ServiceHandleContainer<kom::ProxyHandleType>& container)
-    {
-        // The service level of AUTOSAR Adaptive is not available in iceoryx, instead every publisher and server is
-        // considered as a service. A ara::com binding implementer would typically query the AUTOSAR meta model here, to
-        // find out if all event, fields and methods of a service are available. For the example we assume that the
-        // 'MinimalSkeleton' service is complete when the container contains the four iceoryx services:
-        //
-        // 1.    EventPublisher: MinimalSkeleton, Instance, Event
-        // 2. a) FieldPublisher: MinimalSkeleton, Instance, Field (Publisher)
-        // 2. b) FieldPublisher: MinimalSkeleton, Instance, Field (Server)
-        // 3.    MethodServer:   MinimalSkeleton, Instance, Method
+    bool verifyThatServiceIsComplete(kom::ServiceHandleContainer<kom::ProxyHandleType>& container) noexcept;
 
-        if (container.size() == 4U)
-        {
-            return true;
-        }
-        return false;
-    }
-
-    static void invokeCallback(iox::runtime::ServiceDiscovery*, Runtime* self)
-    {
-        // Has the availability of one of the registered services changed?
-        for (auto& callback : self->m_callbacks)
-        {
-            auto container = self->FindService(std::get<1>(callback).m_serviceIdentifier,
-                                               std::get<1>(callback).m_instanceIdentifier);
-
-            auto numberOfAvailableServicesOnCurrentSearch = container.size();
-            auto& numberOfAvailableServicesOnLastSearch = std::get<2>(callback);
-
-            auto executeCallback = [&]() {
-                (std::get<0>(callback))(container,
-                                        kom::FindServiceHandle({std::get<1>(callback).m_serviceIdentifier,
-                                                                std::get<1>(callback).m_instanceIdentifier}));
-                numberOfAvailableServicesOnLastSearch.emplace(numberOfAvailableServicesOnCurrentSearch);
-            };
-
-            // If the service is available for the first time
-            if (!numberOfAvailableServicesOnLastSearch.has_value() && numberOfAvailableServicesOnCurrentSearch != 0)
-            {
-                executeCallback();
-            }
-
-            // If the service was available before and current number of services has changed
-            if (numberOfAvailableServicesOnLastSearch.has_value()
-                && numberOfAvailableServicesOnLastSearch.value() != numberOfAvailableServicesOnCurrentSearch)
-            {
-                executeCallback();
-            }
-        }
-    }
+    static void invokeCallback(iox::runtime::ServiceDiscovery*, Runtime* self) noexcept;
 
     iox::runtime::ServiceDiscovery m_discovery;
     iox::popo::Listener m_listener;
