+++ iceoryx_examples/automotive_soa/include/topic.hpp
